% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BoxPlot.Calculations.R
\name{BoxPlot.Calculations}
\alias{BoxPlot.Calculations}
\title{Used by analysis funtions to do the calulations for the boxplot graphs}
\usage{
BoxPlot.Calculations(
  f,
  Antibody,
  data.in,
  Opal1,
  Concentration,
  Thresholds,
  x,
  y
)
}
\arguments{
\item{f}{is the number of tiles to use when making boxplots}

\item{Antibody}{is the name of the antibody which the boxplot data is being created for}

\item{data.in}{is a data.frame column of intensity values for the desired Opal}

\item{Opal1}{is the Opal of interest (ie 540)}

\item{Concentration}{is a numeric list of the denominator of all concentration ratios
which a boxplot will be created for (ie 25, 50, 100)}

\item{Thresholds}{is a list of the thresholds at each concentration (ie 5,4,3)}

\item{x}{is the Slide Indicator (ie Tonsil2)}

\item{y}{is the current concentration a boxplot is being made for}
}
\value{
explots a data.frame with the columns: Median, 1st (corresponds to 25th),
 2nd (corresponds to 75th), top.inner.fence, bottom.inner.fence
}
\description{
BoxPlot.Calculations
Created By: Benjamin Green
Last Edited 11/12/2018
}
\details{
This script is desgined to do generate quartiles and fence measurements for creation of boxplots
of decile, thresholded, and all data given IF pixel data
}
