% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.cell.input.vars.R
\name{check.cell.input.vars}
\alias{check.cell.input.vars}
\title{check the cell tab input values
Created By: Benjamin Green
Last Edited 09/03/2020}
\usage{
check.cell.input.vars(out, out.gen, out.chkbx)
}
\arguments{
\item{out}{is the list of variables given by the GUI function}

\item{out.gen}{the values from the general input tab}

\item{out.chckbx}{the values from the analysis check boxes in logical format}
}
\value{
exports multiple variables for use in the main titration codes
}
\description{
Description
This function was designed to check input variables from the cell input
tab for any errors that have occurred in input
}
