% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.chkbx.input.vars.R
\name{check.chkbx.input.vars}
\alias{check.chkbx.input.vars}
\title{evaluate the checkbox input values
Created By: Benjamin Green
Last Edited 09/24/2019}
\usage{
check.chkbx.input.vars(out, a.type)
}
\arguments{
\item{out}{is the list of variables given by the GUI function}

\item{a.type}{is the type of analysis; cell, pixel, or tissue}
}
\value{
exports multiple variables for use in the main titration codes
}
\description{
Description
This function was designed to evaluate the check box input values for later 
use
}
