% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.define.input.paths.R
\name{check.define.input.paths}
\alias{check.define.input.paths}
\title{check and define the input paths for the data
Created By: Benjamin Green
Last Edited 02/09/2021}
\usage{
check.define.input.paths(out.chkbx, out.gen, wd)
}
\arguments{
\item{out.gen}{the values from the general input tab}

\item{out.chckbx}{the values from the analysis check boxes in logical format}
}
\value{
exports the working directory and the folders + check for common errors
}
\description{
Description
This function was designed to check input paths to ensure that they exist and 
that one path exists for each concentration-slide pair (no dup or miss-named
directories are mistaken for a work dir)
}
