% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.EBImage.install.R
\name{check.EBImage.install}
\alias{check.EBImage.install}
\title{check that EBImage installed correctly
Created By: Benjamin Green
Last Edited 09/24/2019}
\usage{
check.EBImage.install(connected.pixels)
}
\arguments{
\item{connected.pixels}{con pixels value}
}
\value{
exports multiple variables for use in the main titration codes
}
\description{
Description
This function was designed to check for the EBImage installation, and install 
any updates necessary. The package does not seem to able to be installed with
these depencies I think since EBImage is a Bioc package and not a CRAN 
package. EBImage is only used when removing objects less than the 
'connected.pixel' values, spefically for the C level bwlabel function it
makes use of.
}
