% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.gen.input.vars.R
\name{check.gen.input.vars}
\alias{check.gen.input.vars}
\title{check the general input tab input values
Created By: Benjamin Green
Last Edited 09/24/2019}
\usage{
check.gen.input.vars(out)
}
\arguments{
\item{out}{is the list of variables given by the GUI function}
}
\value{
exports multiple variables for use in the main titration codes
}
\description{
Description
This function was designed to check input variables from the general input
tab for any errors that have occured in input
}
