% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.threshold.input.vars.R
\name{check.threshold.input.vars}
\alias{check.threshold.input.vars}
\title{check the pixel tab input values
Created By: Benjamin Green
Last Edited 02/09/2021}
\usage{
check.threshold.input.vars(out, out.gen, out.chkbx, ihc.logical)
}
\arguments{
\item{out}{is the list of variables given by the GUI function}

\item{out.gen}{the values from the general input tab}

\item{out.chckbx}{the values from the analysis check boxes in logical format}
}
\value{
exports multiple variables for use in the main titration codes
}
\description{
Description
This function was designed to check input variables from the threshold input
tab for any errors that have occurred in input
}
