% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.vars.R
\name{check.vars}
\alias{check.vars}
\title{Check that the input variables are accurate and get a working directory
CheckVars
Created By: Benjamin Green
Last Edited 09/24/2019}
\usage{
check.vars(out, pb.Object)
}
\arguments{
\item{out}{is the list of variables given by the GUI function}
}
\value{
exports multiple variables for use in the main titration codes
}
\description{
Description
This function was designed to check input variables for any unforseen errors
that may occur
}
