% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CheckVars.R
\name{CheckVars}
\alias{CheckVars}
\title{Check that the input variables are accurate and get a working directory
CheckVars
Created By: Benjamin Green
Last Edited 09/24/2019}
\usage{
CheckVars(out)
}
\arguments{
\item{out}{is the list of vairables given by the GUI function}
}
\value{
exports multiple variables for use in the main titration codes
}
\description{
Description
This function was designed to check input variables for any unforseen errors that may
occur
}
