% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.dir.R
\name{create.dir}
\alias{create.dir}
\title{Create the output directories for either Results.pixels or 
Results.cell recursively checking that the full file tree will be created;
Created By: Benjamin Green;
Last Edited 09/25/2019}
\usage{
create.dir(wd, type, flowout)
}
\arguments{
\item{wd}{the main data root directory}

\item{type}{the second half of the results name (either pixels or cell)}

\item{flowout}{logical for whether or not flow like results will be produced}
}
\description{
Create the output directories for either Results.pixels or 
Results.cell recursively checking that the full file tree will be created
}
