% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.histo.R
\name{create.histo}
\alias{create.histo}
\title{Used by analysis funtions to generate histogram graphs}
\usage{
create.histo(data.in, bin.estimate)
}
\arguments{
\item{data.in}{is a vector of the data of interest}

\item{bin.estimate}{is the estimation of the bin to use if the FD choice fails}
}
\value{
a data.frame with mids, density, and counts columns
}
\description{
create.histo
Created By: Benjamin Green, Charles Roberts
Last Edited 11/12/2018
}
\details{
Takes in a list of values as a data.frame and outputs a histogram caculated based of the Freedman-Diaconis' choice for
bin number if that calculation fails it creates a histogram based off the bin.estimate value
}
