% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.my.theme.R
\name{create.my.theme}
\alias{create.my.theme}
\title{Create a common theme for graphs that is relatable to multiple types of 
graphs
CreateMyTheme
Created By: Benjamin Green
Last Edited 09/25/2019}
\usage{
create.my.theme()
}
\value{
exports the theme and colors for graphs
}
\description{
This function creates a usable theme for graphs so that the theme is common 
the multiple graphs produced
}
