% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateDir.R
\name{CreateDir}
\alias{CreateDir}
\title{Create the output directories for either Results.pixels or 
Results.cell recursively checking that the full file tree will be created;
Created By: Benjamin Green;
Last Edited 09/25/2019}
\usage{
CreateDir(wd, type)
}
\arguments{
\item{wd}{the directory where the results folder belongs}

\item{type}{the second half of the results name (either pixels or cell)}
}
\description{
Create the output directories for either Results.pixels or 
Results.cell recursively checking that the full file tree will be created
}
