% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv.list.R
\name{csv.list}
\alias{csv.list}
\title{Used by analysis functions to read cell seg data text files}
\usage{
csv.list(wd, pattern.in, compartment, phenotype.logical, measurement = "Mean")
}
\arguments{
\item{wd}{is the working directory}

\item{pattern.in}{is the pattern used by R to determine which image or 
images to read in}

\item{compartment}{the compartment of interest to be extracted from the tables
will always extract mean values}

\item{phenotype.logical}{for cell analysis whether the data was phenotyped 
or not}

\item{measurement}{the measurement to extract default is mean}
}
\value{
is a list of two; data.out a column data.frame where each column
holds the pixel intensities for a single image layer, labeled by the inForm
image layer names, and err.val a value indicating an error (1) or not (0)
}
\description{
csv.list;
Created By: Benjamin Green;
Last Edited 11/12/2018
}
\details{
This script is designed to read in the
cell seg text file exported from inForm (R) CellAnaylsis
The function returns a data.frame of n columns, each column in the
data.frame designates a different layer of the image designated
here as DAPI; Flours; AF; however the column/names order may change
depending on the library used to export images from inForm(R)
}
