% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define.image.positivity.R
\name{define.image.positivity}
\alias{define.image.positivity}
\title{Used by analysis functions to calculate to define image positivity using the
thresholds and connected pixel value}
\usage{
define.image.positivity(data.in, threshold, connected.pixels.now)
}
\arguments{
\item{data.in}{is the image for which positivity needs to be defined}

\item{threshold}{is the current threshold}

\item{connected.pixels.now}{is the current connected pixels value}
}
\value{
a list with three data.frames; a sn means, sn medians, and a
fraction of pos
}
\description{
define.image.positivity;
Created By: Benjamin Green;
Last Edited 09/25/2019
}
\details{
Takes in an image, a threshold, and connected pixel values
}
