% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DefineImagePositivity.R
\name{DefineImagePositivity}
\alias{DefineImagePositivity}
\title{Used by analysis funtions to calculate the signal and noise populations}
\usage{
DefineImagePositivity(data.in, threshold, connected.pixels.now)
}
\arguments{
\item{data.in}{is the image for which positivity needs to be defined}

\item{threshold}{is the curren threshold}

\item{connected.pixels.now}{is the current connected pixels value}
}
\value{
a list with three data.frames; a sn means, sn medians, and a fraction of pos
}
\description{
SN.Populations;
Created By: Benjamin Green;
Last Edited 09/25/2019
}
\details{
Takes in an image, a threshold, and connected pixel values
}
