% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doupdate.pgbar.R
\name{doupdate.pgbar}
\alias{doupdate.pgbar}
\title{Update the progress bar in mIFTO;
Created By: Benjamin Green;
Last Edited 09/25/2019}
\usage{
doupdate.pgbar(ii, pb, ll)
}
\arguments{
\item{ii}{level}

\item{pb}{is the progress bar created by the GUI}

\item{ll}{is the label for the progress bar}
}
\value{
updates the progress bar
}
\description{
This function is designed to update the progress bar in mIFTO
}
\details{
It is meant to be run through the RUN.ByImage function
}
