% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fraction.pos.R
\name{fraction.pos}
\alias{fraction.pos}
\title{Used by analysis functions to calculate fraction of positivity}
\usage{
fraction.pos(data.all, data.interest)
}
\arguments{
\item{data.all}{the denominator the division}

\item{data.interest}{the numerator of the division}
}
\value{
a fraction of input 2/ input 1
}
\description{
fraction.pos
Created By: Benjamin Green, Charles Roberts
Last Edited 11/12/2018
}
\details{
This function divides the length of input 2 by the length of input 1
}
