% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.pxp.image.data.R
\name{generate.pxp.image.data}
\alias{generate.pxp.image.data}
\title{Used to create the pixel-by-pixel graph data and return a table;
Created By: Benjamin Green;
Last Edited 08/11/2020}
\usage{
generate.pxp.image.data(
  Concentration,
  x,
  y,
  q,
  Antibody_Opal,
  titration.type.name,
  Thresholds,
  paths,
  connected.pixels,
  flowout,
  Opal1,
  decile.logical,
  threshold.logical
)
}
\arguments{
\item{Concentration}{a numeric vector of concentrations used in the titration}

\item{x}{a unique identifier for the slide to be analyzed}

\item{y}{the numeric of which index from the concentration vector to use}

\item{q}{the image coordinates of the current image as a comma separated pair}

\item{Antibody_Opal}{the paired string for an antibody opal pair, designated
as "AB (Opal NNN)"}

\item{titration.type.name}{the type of titration that was performed
(TSA or Primary)}

\item{Thresholds}{a list of thresholds used for each concentration and slide}

\item{paths}{the paths to the data as a list, with an element for each
concentration}

\item{connected.pixels}{the number of pixels that a pixel must be connected
to for positivity measures}

\item{flowout}{logical for whether or not flow like results will be produced}

\item{Opal1}{the opal value of interest}

\item{decile.logical}{whether or not to run a decile approach analysis}

\item{threshold.logical}{whether or not to run a threshold approach analysis}
}
\value{

}
\description{
This function is desgined to do analysis for IF titration series
in Pixel by Pixel data provding output for each IMAGE individually
grouped by Concentration
}
\details{
It is meant to be run through the RUN function
}
