% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GatherPxPData.R
\name{GeneratePxPData}
\alias{GeneratePxPData}
\title{Used to create the pixel-by-pixel graph data and return a table;
Created By: Benjamin Green;
Last Edited 09/25/2019}
\usage{
GeneratePxPData(Concentration, x, y, q, Antibody_Opal, titration.type.name,
  Protocol, Thresholds, paths, connected.pixels, flowout, Opal1)
}
\description{
This function is desgined to do analysis for IF titration series 
in Pixel by Pixel data provding output for each IMAGE individually 
grouped by Concentration
}
\details{
It is meant to be run through the RUN function
}
