% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IC.Plots.Calculations.R
\name{IC.Plots.Calculations}
\alias{IC.Plots.Calculations}
\title{Usedby analysis functions for violin plot calculations}
\usage{
IC.Plots.Calculations(All.Images, Opal1, Concentration, x, y, colors)
}
\arguments{
\item{All.Images}{the full image vector of a case}

\item{Opal1}{is the Opal of interest (ie 540)}

\item{Concentration}{is a numeric vector of the denominator of all concentration ratios
which a boxplot will be created for (ie 25, 50, 100)}

\item{x}{is the Slide Indicator (ie Tonsil2)}

\item{y}{is the numeric value of the current concentration}

\item{colors}{is a vector of at least the length of the concentration vector}
}
\value{
explots a data.frame with the columns: Median, 1st (corresponds to 25th),
 2nd (corresponds to 75th), top.inner.fence, bottom.inner.fence
}
\description{
IC.Plots.Calculations
Created By: Benjamin Green, Charles Roberts
Last Edited 11/12/2018
}
\details{
This function is desgined to do generate quartiles and fence measurements for creation of boxplots
of decile, thresholded, and all data given IF pixel data; the function;
is also designed to return boxplots of the selected data
}
