% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ihc.generate.pxp.image.data.R
\name{ihc.generate.pxp.image.data}
\alias{ihc.generate.pxp.image.data}
\title{Used to create the pixel-by-pixel graph data and return a table for ihc;
Created By: Benjamin Green;
Last Edited 09/25/2019}
\usage{
ihc.generate.pxp.image.data(
  ihc.path,
  x,
  ihc.Thresholds,
  ihc.connected.pixels,
  z
)
}
\arguments{
\item{ihc.path}{the paths to the data as a list, with an element for each
concentration}

\item{x}{a unique identifier for the slide to be analyzed}

\item{ihc.Thresholds}{a list of thresholds used for each concentration and slide}

\item{ihc.connected.pixels}{the number of pixels that a pixel must be connected
to for positivity measures}

\item{z}{The current image ID}
}
\value{

}
\description{
This function is desgined to do analysis for the ihc in an IF titration series
in Pixel by Pixel data provding output for each IMAGE individually
grouped by Concentration
}
\details{
It is meant to be run through the RUN function
}
