% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ihc.parallel.invoke.gpxp.R
\name{ihc.parallel.invoke.gpxp}
\alias{ihc.parallel.invoke.gpxp}
\title{Used to loop through each image, call the gen func in parallel and return for
IHC
Created By: Benjamin Green;
Last Edited 09/25/2019}
\usage{
ihc.parallel.invoke.gpxp(
  ihc.path,
  x,
  ihc.Image.IDs,
  ihc.Thresholds,
  ihc.connected.pixels,
  cl
)
}
\arguments{
\item{ihc.path}{the paths to the data as a list, with an element for each
concentration}

\item{x}{a unique identifier for the slide to be analyzed}

\item{ihc.Image.IDs}{the image coordinates of the current image as a comma
separated pair}

\item{ihc.Thresholds}{a list of thresholds used for each concentration and slide}

\item{ihc.connected.pixels}{the number of pixels that a pixel must be connected
to for positivity measures}

\item{cl}{cluster object}
}
\value{

}
\description{
This function is desgined launch the parallelization of the ihc.gpxp function in
the mIFTO code. It is used to create a local environment with only necessary
data for the cluster objects.
}
\details{
It is meant to be run through the RUN function
}
