% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m.grid.arrange.R
\name{m.grid.arrange}
\alias{m.grid.arrange}
\title{Used to create the figures
Created By: Benjamin Green
Last Edited 06/04/2020}
\usage{
m.grid.arrange(p, lbl, lbl2, opt, st.pg, total.pgs)
}
\arguments{
\item{p}{a list of plots}

\item{lbl}{the titles for each figure}

\item{lbl2}{the subtitles for each figure}
}
\value{
returns the layout object
}
\description{
This function is desgined to build the arrangement of plots for the mIFTO. 
Four plots will be displayed on each page returned.
}
