% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m.validation.R
\name{m.validation}
\alias{m.validation}
\title{validate that all variables have been populated before moving on
Created By: Benjamin Green;
Last Edited 06/12/2020}
\usage{
m.validation(input)
}
\arguments{
\item{input}{input from UI}
}
\value{
app.e servers as the error message value, depending on the value the 
server side will know how to respond to errors.
}
\description{
Sets up the formats for all buttons in the UI
}
