% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.and.plot.decile.graphs.R
\name{map.and.plot.decile.graphs}
\alias{map.and.plot.decile.graphs}
\title{Used by PxP script to write out the decile type graphs}
\usage{
map.and.plot.decile.graphs(
  wd,
  Antibody_Opal,
  Antibody,
  Slide_Descript,
  Concentration,
  Tables,
  theme1,
  con_type,
  colors,
  Antibody_Opal.snratio,
  Antibody_Opal.ttest,
  pb.Object
)
}
\arguments{
\item{wd}{the main data root directory}

\item{Antibody_Opal}{the paired string for an antibody opal pair, designated
as "AB (Opal NNN)"}

\item{Antibody}{the string for an antibody}

\item{Concentration}{a numeric vector of concentrations used in the titration}

\item{Tables}{the table of statistics gathered by PxP}

\item{theme1}{graphing theme}

\item{con_type}{the type of concentration vector to use factor or numeric}

\item{colors}{the color vectors for the t test and histograms}

\item{Antibody_Opal.snratio}{Antibody Opal string for sn ratio graphs}

\item{Antibody_Opal.ttest}{Antibody Opal string for ttest graphs}

\item{pb.Object}{progress bar object}

\item{Slide_Desctipt}{a unique identifier for each slide to be analyzed}
}
\value{
exports the fraction spreadsheets
}
\description{
Created By: Benjamin Green
Last Edited 07/30/2020
}
\details{
Designed to write out the fraction of positivity data for the PxP script

It is meant to be run through the PixelbyPixel function
}
