% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.and.write.histograms.R
\name{map.and.write.histograms}
\alias{map.and.write.histograms}
\title{Used by PxP script to plot out the histograms}
\usage{
map.and.write.histograms(
  wd,
  Antibody_Opal,
  Slide_Descript,
  Concentration,
  Thresholds,
  table_in,
  theme1,
  colors
)
}
\arguments{
\item{wd}{the main data root directory}

\item{Antibody_Opal}{the paired string for an antibody opal pair,
designated as "AB (Opal NNN)"}

\item{Concentration}{a numeric vector of concentrations used in the titration}

\item{Thresholds}{The given threshold vector}

\item{table_in}{the table of statistics gathered by PxP}

\item{theme1}{the theme for the graphs}

\item{colors}{the color vectors for the t test and histograms}

\item{Slide_Desctipt}{a unique identifier for each slide to be analyzed}
}
\value{
exports various data and histogram plots
}
\description{
Created By: Benjamin Green, Charles Roberts
Last Edited 05/21/2020
}
\details{
Designed to output the histogram plots and data

It is meant to be run through the PixelbyPixel function
}
