% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.ihc.comp.plots.R
\name{map.ihc.comp.plots}
\alias{map.ihc.comp.plots}
\title{Used by PxP script to write out the fraction of positivity graphs}
\usage{
map.ihc.comp.plots(
  wd,
  Antibody_Opal,
  Slide_Descript,
  Concentration,
  tbl.long,
  theme1
)
}
\arguments{
\item{wd}{the main data root directory}

\item{Antibody_Opal}{the paired string for an antibody opal pair, designated
as "AB (Opal NNN)"}

\item{Concentration}{a numeric vector of concentrations used in the titration}

\item{tbl.long}{the long type version of the fractions table}

\item{theme1}{graphing theme}

\item{Slide_Desctipt}{a unique identifier for each slide to be analyzed}
}
\value{
exports the fraction spreadsheets
}
\description{
Created By: Benjamin Green
Last Edited 07/30/2020
}
\details{
Designed to write out the fraction of positivity graph for comparing different
dilutions to IHC

It is meant to be run through the PixelbyPixel function
}
