% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_snratio_plots.R
\name{map_snratio_plots}
\alias{map_snratio_plots}
\title{Used by PxP script to plot out the snratio plots}
\usage{
map_snratio_plots(
  wd,
  Antibody_Opal,
  Slide_Descript,
  Concentration,
  tables_in,
  theme1
)
}
\arguments{
\item{wd}{the main data root directory}

\item{Antibody_Opal}{the paired string for an antibody opal pair, designated as 
"AB (Opal NNN)"}

\item{Concentration}{a numeric vector of concentrations used in the titration}

\item{theme1}{the theme for the graphs}

\item{Slide_Desctipt}{a unique identifier for each slide to be analyzed}

\item{Tables}{the table of statistics gathered by PxP}
}
\value{
exports a ggplot object to be printed for viewing
}
\description{
Created By: Benjamin Green, Charles Roberts
Last Edited 05/21/2020
}
\details{
Designed to output the sn ratio plots as a ggplot object for the PxP script

It is meant to be run through the PixelbyPixel function
}
