% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_ttest_plots.R
\name{map_ttest_plots}
\alias{map_ttest_plots}
\title{Used by PxP script to plot out the ttests}
\usage{
map_ttest_plots(
  wd,
  Antibody_Opal,
  Slide_Descript,
  Concentration,
  Tables,
  theme1,
  colors
)
}
\arguments{
\item{wd}{the main data root directory}

\item{Antibody_Opal}{the paired string for an antibody opal pair, designated as 
"AB (Opal NNN)"}

\item{Concentration}{a numeric vector of concentrations used in the titration}

\item{Tables}{the table of statistics gathered by PxP}

\item{theme1}{the theme for the graphs}

\item{colors}{the color vectors for the t test and histograms}

\item{Slide_Desctipt}{a unique identifier for each slide to be analyzed}
}
\value{
exports a ggplot object to be printed for viewing
}
\description{
Created By: Benjamin Green, Charles Roberts
Last Edited 05/21/2020
}
\details{
Designed to output the ttest plots as a ggplot object for the PxP script

It is meant to be run through the PixelbyPixel function
}
