% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PixelbyPixel.BySlide.R
\name{PixelbyPixel.BySlide}
\alias{PixelbyPixel.BySlide}
\title{Used by RUN.BySlide to do Pixel by Pixel Analysis on images for IF titrations}
\usage{
PixelbyPixel.BySlide(out, pb)
}
\arguments{
\item{out}{is the list of vairables given by the GUI function}

\item{pb}{is the progress bar created by the GUI}
}
\value{
exports a variety of graphs displayed in the documentation
 Such as SNRatio graphs, t statisitics and graphs,
 histograms of the log intensity profiles, boxplots and
 violin plots of the profiles for images,
 positivity measures given thresholds
}
\description{
Created By: Benjamin Green, Charles Roberts
Last Edited 11/12/2018
}
\details{
This function is desgined to do analysis for IF titration series in Pixel by Pixel
data provding output for each SLIDE individually grouped by Concentration

it is meant to be run through the RUN.BySlide function

decile data will always be outputed; (or 1/100th depending if
'sparse' option is choose in the GUI) if threshold information is
filled out in the GUI; threshold analysis will be run
}
