% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pixelbypixel.R
\name{pixelbypixel}
\alias{pixelbypixel}
\title{Used by RUN to do Pixel by Pixel Analysis on individual images for
IF titrations;
Created By: Benjamin Green, Charles Roberts;
Last Edited 09/25/2019}
\usage{
pixelbypixel(out, pb.Object)
}
\arguments{
\item{out}{is the list of variables given by the GUI function}

\item{pb}{is the progress bar created by the GUI}
}
\value{
exports a variety of graphs displayed in the documentation
 Such as SNRatio graphs, t statisitics and graphs,
 histograms of the log intensity profiles
 for images, positivity measures given thresholds
}
\description{
This function is desgined to do analysis for IF titration series
in Pixel by Pixel data provding output for each IMAGE individually
grouped by Concentration
}
\details{
It is meant to be run through the RUN.ByImage function

decile data will always be outputed; (or 1/100th depending if
'sparse' option is choose in the GUI) if threshold information is
filled out in the GUI; threshold analysis will be run
}
