% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populate_mIFTO_tables.R
\name{populate_mIFTO_tables}
\alias{populate_mIFTO_tables}
\title{Used to loop through each image, call the generate func and map back to Tables;
Created By: Benjamin Green;
Last Edited 09/25/2019}
\usage{
populate_mIFTO_tables(
  Slide_Descript,
  Tables,
  Image.IDs,
  Concentration,
  Antibody_Opal,
  Thresholds,
  Opal1,
  table.names,
  outchecked,
  pbi1,
  ii,
  pb
)
}
\arguments{
\item{Tables}{the table of statistics gathered by PxP}

\item{Image.IDs}{the list of image coordinates index as slide then concentration}

\item{Concentration}{a numeric vector of concentrations used in the titration}

\item{Antibody_Opal}{the paired string for an antibody opal pair, designated as 
"AB (Opal NNN)"}

\item{Thresholds}{a list of thresholds used for each concentration and slide}

\item{Opal1}{the opal value of interest}

\item{outchecked}{the checked version (by mIFTO::CheckVars) of the UI output}

\item{Slide_Desctipt}{a unique identifier for each slide to be analyzed}
}
\value{

}
\description{
This function is desgined to do analysis for IF titration series 
in Pixel by Pixel data provding output for each IMAGE individually 
grouped by Concentration
}
\details{
It is meant to be run through the RUN function
}
