% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preallocate.tables.R
\name{preallocate.tables}
\alias{preallocate.tables}
\title{Preallocate the tables for the results values;
Created By: Benjamin Green;
Last Edited 09/25/2019}
\usage{
preallocate.tables(
  Slide_Descript,
  Concentration,
  titration.type.name,
  table.names,
  paths
)
}
\arguments{
\item{Concentration}{a numeric vector of concentrations used in the titration}

\item{titration.type.name}{the titration type for a given dilution set (Primary or TSA)}

\item{table.names}{the table names for whole slide names}

\item{paths}{the data paths, one data path for each concentration}

\item{Slide_Desctipt}{a unique identifier for each slide to be analyzed}
}
\value{
exports the Tables list and the Image.IDs sublists
}
\description{
Preallocate the tables for the results values. All graphs will be 
generated based off of values stored in the Tables list and all 
ImageIDs (x and y coordinates) will be added to the Image.IDs 
list later. This function prepopulates both of these tables 
with separate subvectors for each concentration - slide pair
}
