% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bonzai.R
\name{RUN}
\alias{RUN}
\title{Main funtion to start analysis on a by image basis. Creates the gui and 
calls other functions within the package for processing}
\usage{
RUN()
}
\value{
a variety of ouput data will be displayed in a Results subdirectory in the selected data directory
}
\description{
RUN.ByImage
Created By: Benjamin Green, Charles Roberts
Last Edited 11/12/2018
}
\details{
This function is designed to create the GUI for the TitrationScript
Here you should input the parameters for the given dilution series of interest
All of the input parameters are described in additional documentation for TitrationScript
Sections are as follows;
1: General Input
2: Information for Cell Segmented Based Data Analysis
3: Information for Pixel Based Data Analysis

Run the script for cellbycell data by filling out sections 1 and 2;
then click 'Run for Cell Segmented Based Analysis'
Run the script for pixelbypixel data by filling out section 1 and 3;
then click 'Run for Pixel Based Analysis'

for both sections after run is selected a windows explorer will open; direct 
the explorer to the desired file ouput
}
