% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.titration.analysis.R
\name{run.titration.analysis}
\alias{run.titration.analysis}
\title{The main workflow wrapper for analysis; reads files, collects data, and 
produces output
Created By: Benjamin Green;
Last Edited 09/02/2020}
\usage{
run.titration.analysis(out, pb.Object, a.type)
}
\arguments{
\item{out}{is the list of variables given by the GUI function}

\item{a.type}{is the type of analysis; cell, pixel, or tissue}

\item{pb}{is the progress bar created by the GUI}
}
\value{
exports a variety of graphs displayed in the documentation
 Such as SNRatio graphs, t statisitics and graphs,
 histograms of the log intensity profiles
 for images, positivity measures given thresholds
}
\description{
This function is desgined to do analysis for IF titration series
in cell, tissue, or pixel data provding output grouped by slide and 
concentrations; both in an image and a cell setting
}
\details{
It is meant to be run from the mIFTOapp

decile data will always be outputed; (or 1/100th depending if
'sparse' option is choose in the GUI) if threshold information is
filled out in the GUI; threshold analysis will be run
}
