% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.side.R
\name{server.side}
\alias{server.side}
\title{Used by mIFTO to run the main App
Created By: Benjamin Green;
Last Edited 06/12/2020}
\usage{
server.side(input, output, session)
}
\arguments{
\item{input}{input from UI}

\item{output}{output from computation}
}
\value{
updates UI or runs relevent functions
}
\description{
Sets up the formats for all buttons in the UI
}
