% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SN.Ratio.Calculations.R
\name{SN.Ratio.Calculations}
\alias{SN.Ratio.Calculations}
\title{Used by analysis funtions to calculate the signal to noise ratios}
\usage{
SN.Ratio.Calculations(positivity.data, Concentration, x, q)
}
\arguments{
\item{positivity.data}{a list of four images, positive, 
negative values then positive, negative binary image masks}

\item{Concentration}{is a numeric vector of the denominator of all concentration ratios
which a boxplot will be created for (ie 25, 50, 100)}

\item{x}{is the Slide Indicator (ie Tonsil2)}

\item{q}{is the imageid}
}
\value{
a list with three data.frames; a sn means, sn medians, and a fraction of pos
}
\description{
SN.Ratio.Calculations
Created By: Benjamin Green, Charles Roberts
Last Edited 11/12/2018
}
\details{
Takes in the data and some constants from the PixelbyPixel script and returns a SN ratio for means and by medians
}
