% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sn.ratio.theme.R
\name{sn.ratio.theme}
\alias{sn.ratio.theme}
\title{Used by PxP script to create the sn_ratio_theme}
\usage{
sn.ratio.theme(tbl, Concentration, titl, xtitl, ytitl, Max, theme1)
}
\arguments{
\item{tbl}{the table to be graphed}

\item{Concentration}{a numeric vector of concentrations used in the titration}

\item{titl}{the title of the graph}

\item{xtitl}{the x axis label}

\item{ytitl}{the y axis label}

\item{Max}{max value}

\item{theme1}{the additional theme settings for the graphs defined by PxP}
}
\value{
exports the fraction spreadsheets
}
\description{
Created By: Benjamin Green, Charles Roberts
Last Edited 05/21/2020
}
\details{
Designed to export a plot object with the specified settings for the sn
ratio graphs

It is meant to be run through the PixelbyPixel function
}
