% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sn_ratio_theme.R
\name{sn_ratio_theme}
\alias{sn_ratio_theme}
\title{Used by PxP script to create the sn_ratio_theme}
\usage{
sn_ratio_theme(tbl, titl, xtitl, ytitl, theme1)
}
\arguments{
\item{tbl}{the table to be graphed}

\item{xtitl}{the x axis label}

\item{ytitl}{the y axis label}

\item{theme1}{the additional theme settings for the graphs defined by PxP}

\item{the}{title of the graph}
}
\value{
exports the fraction spreadsheets
}
\description{
Created By: Benjamin Green, Charles Roberts
Last Edited 05/21/2020
}
\details{
Designed to export a plot object with the specified settings for the sn
ratio graphs

It is meant to be run through the PixelbyPixel function
}
