% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thresholding.medians.R
\name{thresholding.medians}
\alias{thresholding.medians}
\title{Used by analysis functions to calculate the median signal to noise ratios}
\usage{
thresholding.medians(
  data.in,
  threshold,
  Slide.ID,
  Concentration,
  connected.pixels
)
}
\arguments{
\item{data.in}{a single column data table with the values of interest}

\item{threshold}{a single value to divide the population, everything equal to and above is positive}

\item{Slide.ID}{a slide distinguisher}

\item{Concentration}{is the concentration of the choosen AB}
}
\value{
a data.frame with Signal, Noise, S/N, Slide.ID, Concentration
}
\description{
thresholding.medians
Created By: Benjamin Green
Last Edited 11/12/2018
}
\details{
This function divides the given data at the threshold given and calculates a median for the above and below values
The function then calculates and returns a signal to noise calculation
}
