% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thresholding.Noise.R
\name{thresholding.Noise}
\alias{thresholding.Noise}
\title{Used by analysis funstions to return all noise values above given threshold}
\usage{
thresholding.Noise(data.in, threshold)
}
\arguments{
\item{data.in}{a single column data table with the values of interest}

\item{threshold}{a single value to divide the population, everything equal to and above is positive}
}
\value{
a data.frame with Noise intensity values
}
\description{
thresholding.Noise
Created By: Benjamin Green, Charles Roberts
Last Edited 11/12/2018
}
\details{
This function divides the given data at the threshold given and provides all values above the threshold
}
