% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/titration_analysis.R
\name{titation_analysis}
\alias{titation_analysis}
\title{Used by mIFTOapp to do titration_analysis on individual images for
IF titrations;
Created By: Benjamin Green, Charles Roberts;
Last Edited 02/11/2021}
\usage{
titation_analysis(out, pb.Object, a.type)
}
\arguments{
\item{out}{is the list of variables given by the GUI function}

\item{pb}{is the progress bar created by the GUI}
}
\value{
exports a variety of graphs and tables, for more information please 
reference the readme file in the github repo
}
\description{
This function is designed to aid in the analysis of an IF titration series. 
The code collects data (pixel, cell seg, or tissue segmented data) from the
inForm output and, using specified input to the mIFTOapp UI, takes particular
measurements, performs statistical testing and outputs data tables and graphical
pdfs. In this code each image is treated as a separate data point taken for 
each slide/ concentration pair. Data is grouped and graphed accordingly.
}
\details{
The code is meant to be run through the mIFTOapp, so particular input to the
'out' variable is not specified in great detail. Additional information can 
be found in the mIFTO readme file in the github repo at AstropathJHU/mIFTO
}
