% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.formats.R
\name{ui.formats}
\alias{ui.formats}
\title{Used by mIFTO to run the main App
Created By: Benjamin Green;
Last Edited 06/12/2020}
\usage{
ui.formats(w, opts)
}
\arguments{
\item{w}{the width of the UI}

\item{opts}{if value is 1 the full format list is returned otherwise only 
commontextstyle and fineprintstyle are returned}
}
\value{
a list of different fonts and formatting styles which will be 
referenced later on
}
\description{
Sets up the formats for all buttons in the UI
}
