% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.map.R
\name{ui.map}
\alias{ui.map}
\title{Used by mIFTO to create the UI map
Created By: Benjamin Green;
Last Edited 06/12/2020}
\usage{
ui.map()
}
\value{
returns the UIMap
}
\description{
This function is designed to create the UI layout for the main mIFTO app
}
