% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_pgbar.R
\name{update_pgbar}
\alias{update_pgbar}
\title{Update the progress bar in mIFTO;
Created By: Benjamin Green;
Last Edited 09/25/2019}
\usage{
update_pgbar(ii, pb, ll)
}
\arguments{
\item{ii}{level}

\item{pb}{is the progress bar created by the GUI}

\item{ll}{is the label for the progress bar}
}
\value{
updates the progress bar
}
\description{
This function is desgined to update the progress bar in mIFTO
}
\details{
It is meant to be run through the RUN.ByImage function
}
