% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Welch.t.test.R
\name{Welch.t.test}
\alias{Welch.t.test}
\title{Used by analysis functions to perform Welch T-Test on input data vectors}
\usage{
Welch.t.test(x1, x2)
}
\arguments{
\item{x1}{the first vector of data}

\item{x2}{the second vector of data to compare}
}
\value{
A t statistic value for the welch t test
}
\description{
Welch t test
Created By: Benjamin Green, Charles Roberts
Last Edited 11/12/2018
}
\details{
This function performs a Welch t test on the data vectors provided
}
