% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_fracs.R
\name{write_fracs}
\alias{write_fracs}
\title{Used by PxP script to write out the fraction of positivity data}
\usage{
write_fracs(wd, Antibody_Opal, Slide_Descript, Concentration, Tables, IHC)
}
\arguments{
\item{wd}{the main data root directory}

\item{Antibody_Opal}{the paired string for an antibody opal pair, designated as 
"AB (Opal NNN)"}

\item{Concentration}{a numeric vector of concentrations used in the titration}

\item{Tables}{the table of statistics gathered by PxP}

\item{IHC}{whether or not an IHC was done on this slide}

\item{Slide_Desctipt}{a unique identifier for each slide to be analyzed}
}
\value{
exports the fraction spreadsheets
}
\description{
Created By: Benjamin Green, Charles Roberts
Last Edited 05/21/2020
}
\details{
Designed to write out the fraction of positivity data for the PxP script

It is meant to be run through the PixelbyPixel function
}
