% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Histogram.Calculations.R
\name{Histogram.Calculations}
\alias{Histogram.Calculations}
\title{Used by analysis funtions to do the calculations for the histogram graphs}
\usage{
Histogram.Calculations(data.in, Concentration, x, q)
}
\arguments{
\item{data.in}{is the image of interest}

\item{Concentration}{is a numeric vector of the denominator of all concentration ratios
which a boxplot will be created for (ie 25, 50, 100)}

\item{x}{is the slide id}

\item{q}{is the image id}
}
\value{
a list with two data.frames; one with a histogram for shift factor of log(x+1) and one with a shift factor of log(x+.001)
the data.frames hold four columns three for the create.hist output and one for the concentration of interest
}
\description{
Histogram.Calculations
Created By: Benjamin Green, Charles Roberts
Last Edited 11/12/2018
}
\details{
Takes in the data and some constants from the PixelbyPixel script and returns a list with two data.frames
one for a histogram of the log(x+1) of the data and one for a histogram of log(x+.001) of the data
data.in is an col data.frame; Concentration: [25,50,100]; Opal1 540; Thresholds: [2,5,8];
to run for Tonsil2, 1to25 with threshold of 2 --> x: Tonsil2; y: 1
}
