% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IC.Plots.Calculations.R
\name{IC.Plots.Calculations}
\alias{IC.Plots.Calculations}
\title{Usedby analysis functions for violin plot calculations}
\usage{
IC.Plots.Calculations(
  f,
  data.in,
  Opal1,
  Concentration,
  Thresholds,
  x,
  y,
  colors
)
}
\arguments{
\item{f}{is the number of tiles to use when making boxplots}

\item{data.in}{is a data.frame column of intensity values for the desired Opal}

\item{Opal1}{is the Opal of interest (ie 540)}

\item{Concentration}{is a numeric vector of the denominator of all concentration ratios
which a boxplot will be created for (ie 25, 50, 100)}

\item{Thresholds}{is a list of the thresholds at each concentration (ie 5,4,3)}

\item{x}{is the Slide Indicator (ie Tonsil2)}

\item{y}{is the numeric value of the current concentration}

\item{colors}{is a vector of at least the length of the concentration vector}

\item{Antibody}{is the name of the antibody which the boxplot data is being created for}
}
\value{
explots a data.frame with the columns: Median, 1st (corresponds to 25th),
 2nd (corresponds to 75th), top.inner.fence, bottom.inner.fence
}
\description{
IC.Plots.Calculations
Created By: Benjamin Green, Charles Roberts
Last Edited 11/12/2018
}
\details{
This function is desgined to do generate quartiles and fence measurements for creation of boxplots
of decile, thresholded, and all data given IF pixel data; the function;
is also designed to return boxplots of the selected data
}
