% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IC.Plots.Calculations.R
\name{ic.plots.calculations}
\alias{ic.plots.calculations}
\title{Usedby analysis functions for violin plot calculations}
\usage{
ic.plots.calculations(All.Images, Opal1, Concentration, x, y, m.opt)
}
\arguments{
\item{All.Images}{the full image vector of a case}

\item{Opal1}{is the Opal of interest (ie 540)}

\item{Concentration}{is a numeric vector of the denominator of all
concentration ratios which a boxplot will be created for (ie 25, 50, 100)}

\item{x}{is the Slide Indicator (ie Tonsil2)}

\item{y}{is the numeric value of the current concentration}

\item{m.opt}{optionally export the percentile boxplots between 10% and 2%,
always export the signal and noise boxplots}
}
\value{
explots a data.frame with the columns: Median, 1st
(corresponds to 25th), 2nd (corresponds to 75th), top.inner.fence,
bottom.inner.fence
}
\description{
IC.Plots.Calculations
Created By: Benjamin Green, Charles Roberts
Last Edited 11/12/2018
}
\details{
This function is desgined to do generate quartiles and fence measurements
for creation of boxplots of decile, thresholded, and all data given IF
pixel data; the function; is also designed to return boxplots of the
selected data
}
