% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populate.tables.R
\name{populate.tables}
\alias{populate.tables}
\title{Used to loop through each image, call the generate function and map back to
Tables;
Created By: Benjamin Green;
Last Edited 05/29/2020}
\usage{
populate.tables(a.type, oc, pb.count, pb.Object)
}
\arguments{
\item{a.type}{is the type of analysis; cell, pixel, or tissue}

\item{oc}{a collection of verified and pre organized variables for use, including:
 - Slide_Descript a unique identifier for each slide to be analyzed
 - Concentration a numeric vector of concentrations used in the titration
 - Antibody_Opal the paired string for an antibody fluor pair, designated
 as "AB ( NNN)"
 - Thresholds a list of thresholds used for each concentration and slide
 - Opal1 the opal value of interest
 - flowout logical for whether or not flow like results will be produced
(1 for produce, 0 for don't)
 - Protocol the protocol type (7color or 9color)
 - paths the data paths, one data path for each concentration
 - titration.type.name the titration type for a given dilution set
(Primary or TSA)
 - decile.logical whether or not to run a decile approach analysis
 - step.value the number of tiles to divide the data into for decile 
approach
 - connected.pixels the number of pixels that a pixel must be connected
to for positivity measures
 - threshold.logical whether or not to run a threshold approach analysis
 - compartment cell compartment for cell analysis
 - phenotype.logical for cell analysis whether the data was phenotyped 
or not
 - pheno.antibody the name of the positive phenotype in cell analysis}

\item{pb.count}{current count for progress bar}

\item{pb.Object}{progress bar object}
}
\value{

}
\description{
This function is designed to do analysis for IF titration series
in Pixel by Pixel data providing output for each IMAGE individually
grouped by Concentration
}
\details{
It is meant to be run through the RUN function
}
