% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/T.Test.Calculations.R
\name{T.Test.Calculations}
\alias{T.Test.Calculations}
\title{Used by analysis functions to do the T-Test calcualtions}
\usage{
T.Test.Calculations(positivity.data, Concentration, x, y, q)
}
\arguments{
\item{positivity.data}{a list of four images, positive, 
negative values then positive, negative binary image masks}

\item{Concentration}{is a numeric vector of the denominator of all concentration ratios
which a boxplot will be created for (ie 25, 50, 100)}

\item{y}{is the numeric value of the current concentration}
}
\value{
a list with two data.frames; one with a t test for shift factor of log(x+1) and one with a shift factor of log(x+.001)
the data.frames hold two columns one for the t test value and one for the concentration of interest
}
\description{
T.Test.Calculations;
Created By: Benjamin Green;
Last Edited 09/27/2019
}
\details{
Takes in the data and some constants from the PixelbyPixel script and returns a T test for postive and negative
Signal defined by the threshold
}
